Fix an async bug with the autoconnect command.
We must be sure that the server is ready before that the client sends
the autoconnect command. Otherwise, the command can be lost and the
client will wait forever.

When the server is ready, it sends an heartbeat every 500ms. The client
looks for a first heartbeat, then it sends the autoconnect command.
Then we are sure that the server has registered its autoconnect
handler.